+++ CHANGELOG.md
@@ -7,6 +7,7 @@ See `UPGRADE.md` for instructions on upgrading between versions.
 * Live preview: Now sends entry diffs on update to save some bandwidth
 * Live preview: Send base64 of images on entry creation
 * Villain: Added telemetry for `parse_and_render`
+* Try to rotate images by EXIF info on upload
 
 
 ## 0.48.0